+++ lib/src/ui/View/Emojis/emojis_variations.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_giphy_picker/giphy_api.dart';
+import 'package:flutter_giphy_picker/src/ui/Functions/get_preview_url.dart';
 import 'package:flutter_giphy_picker/src/ui/Functions/parse_exception.dart';
 import 'package:flutter_giphy_picker/src/ui/View/Emojis/select_emoji.dart';
 import 'package:flutter_giphy_picker/src/ui/giphy_config.dart';
@@ -107,8 +108,10 @@ class _EmojisVariationsView extends State<EmojisVariationsView> {
                                 ),
                               )
                             : Image.network(
-                                emojis[index].images.previewGif?.url ??
-                                    emojis[index].images.original!.webp!,
+                                getGifPreviewUrlForQuality(
+                                  gif: emojis[index],
+                                  quality: widget.config.previewQuality,
+                                ),
                                 fit: BoxFit.contain,
                               ),
                       );