+++ lib/src/ui/View/Emojis/select_emoji.dart
@@ -101,6 +101,7 @@ class _SelectEmojiView extends State<SelectEmojiView> {
               style: ElevatedButton.styleFrom(
                 fixedSize: Size(MediaQuery.of(context).size.width / 3, 30),
                 elevation: 5,
+                padding: const EdgeInsets.all(8),
               ),
               child: Text(
                 translations(context)?.selectEmoji ?? "Select this Emoji",