+++ lib/src/ui/View/Gifs/gifs.dart
@@ -2,6 +2,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_giphy_picker/giphy_api.dart';
 import 'package:flutter_giphy_picker/giphy_ui.dart';
+import 'package:flutter_giphy_picker/src/ui/Functions/get_preview_url.dart';
 import 'package:flutter_giphy_picker/src/ui/Functions/parse_exception.dart';
 import 'package:flutter_giphy_picker/src/ui/Platforms/platforms_stub.dart';
 import 'package:flutter_giphy_picker/src/ui/View/Gifs/categories.dart';
@@ -170,7 +171,10 @@ class _GifsView extends State<GifsView> {
                               );
                             },
                             child: Image.network(
-                              gifs[index].images.previewGif!.url,
+                              getGifPreviewUrlForQuality(
+                                gif: gifs[index],
+                                quality: widget.config.previewQuality,
+                              ),
                               fit: BoxFit.contain,
                             ),
                           );